Add support for dev versions
A new Version model was added to keep all the formats created by Composer for a
given version (ie the name used for parsing, the pretty format and a full pretty
format).

This allows to easily detect dev versions and also to keep a reference of the
short commit hash.

Fixes #3